// proto/data_payload_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file proto/data_payload.proto (package proto.data_payload, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/dist/cjs/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/dist/cjs/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file proto/data_payload.proto.
 */
export const file_proto_data_payload: GenFile = /*@__PURE__*/
  fileDesc("Chhwcm90by9kYXRhX3BheWxvYWQucHJvdG8SEnByb3RvLmRhdGFfcGF5bG9hZCKIAQoERGF0YRITCgtkZXZpY2VfdXVpZBgBIAEoCRITCgt0ZW1wZXJhdHVyZRgCIAEoAhIQCghodW1pZGl0eRgDIAEoAhINCgVsaWdodBgEIAEoAhIVCg1zb2lsX21vaXN0dXJlGAUgASgCEgsKA2NvMhgGIAEoAhIRCgl0aW1lc3RhbXAYByABKAEiOQoLRGF0YVBheWxvYWQSKgoIZGF0YXNldHMYASADKAsyGC5wcm90by5kYXRhX3BheWxvYWQuRGF0YWIGcHJvdG8z");

/**
 * @generated from message proto.data_payload.Data
 */
export type Data = Message<"proto.data_payload.Data"> & {
  /**
   * @generated from field: string device_uuid = 1;
   */
  deviceUuid: string;

  /**
   * @generated from field: float temperature = 2;
   */
  temperature: number;

  /**
   * @generated from field: float humidity = 3;
   */
  humidity: number;

  /**
   * @generated from field: float light = 4;
   */
  light: number;

  /**
   * @generated from field: float soil_moisture = 5;
   */
  soilMoisture: number;

  /**
   * @generated from field: float co2 = 6;
   */
  co2: number;

  /**
   * @generated from field: double timestamp = 7;
   */
  timestamp: number;
};

/**
 * Describes the message proto.data_payload.Data.
 * Use `create(DataSchema)` to create a new message.
 */
export const DataSchema: GenMessage<Data> = /*@__PURE__*/
  messageDesc(file_proto_data_payload, 0);

/**
 * @generated from message proto.data_payload.DataPayload
 */
export type DataPayload = Message<"proto.data_payload.DataPayload"> & {
  /**
   * @generated from field: repeated proto.data_payload.Data datasets = 1;
   */
  datasets: Data[];
};

/**
 * Describes the message proto.data_payload.DataPayload.
 * Use `create(DataPayloadSchema)` to create a new message.
 */
export const DataPayloadSchema: GenMessage<DataPayload> = /*@__PURE__*/
  messageDesc(file_proto_data_payload, 1);